2319: Rename file `PreVerifiedEntitlements.ts` to lowercase
It does not contain a component
